+++ src/Vela.jl
@@ -66,6 +66,11 @@ include("likelihood/gls_ecorr_likelihood.jl")
 include("likelihood/posterior.jl")
 include("pulsar/pulsar.jl")
 include("readwrite/readwrite.jl")
+include("spna/resids.jl")
+include("spna/white_noise.jl")
+include("spna/spnautils.jl")
+include("spna/spna.jl")
+include("spna/lnlike.jl")
 
 pkg_version() = string(PkgVersion.Version(Vela))